HBX-1945: Move class DefaultDatabaseCollector from ' org.hibernate.tool.internal.metadata' to 'org.hibernate.tool.internal.reveng'
